deal with warning

+++ Connection.cpp
@@ -47,7 +47,7 @@ void Connection::start_loading() {
     if (loader_issued >= options.records) break;
     char key[256];
     int index = lrand48() % (1024 * 1024);
-    snprintf(key, 256, "%0*" PRIu64, 30, loader_issued);
+    snprintf(key, 256, "%0*" PRIu64, 30, (unsigned long)loader_issued);
     issue_set(key, &random_char[index], 200);
     loader_issued++;
   }
@@ -55,7 +55,7 @@ void Connection::start_loading() {
 
 void Connection::issue_set_or_get(double now) {
   char key[256];
-  snprintf(key, 256, "%0*" PRIu64, 30, loader_issued);
+  snprintf(key, 256, "%0*" PRIu64, 30, (unsigned long)loader_issued);
 
   if (drand48() < options.update) {
     int index = lrand48() % (1024 * 1024);
@@ -198,7 +198,7 @@ void Connection::read_callback() {
           if (loader_issued >= options.records) break;
           char key[256];
           int index = lrand48() % (1024 * 1024);
-          snprintf(key, 256, "%0*" PRIu64, 30, loader_issued);
+          snprintf(key, 256, "%0*" PRIu64, 30, (unsigned long)loader_issued);
           issue_set(key, &random_char[index], 200);
           loader_issued++;
         }
